+++ cypress/e2e/visualRegressions.cy.ts
@@ -15,7 +15,7 @@ const FIGURES_DATA = [
     { name: "parallelplot", data: parallelPlotData, threshold: 0.05 },
     { name: "plotscatter", data: plotScatterData, threshold: 0.05 },
     { name: "primitivegroupcontainer", data: primitiveGroupContainerData, threshold: 0.05 },
-    { name: "scattermatrix", data: scattermatrixData, threshold: 0.07 },
+    { name: "scattermatrix", data: scattermatrixData, threshold: 0.08 },
     { name: "simpleshapes", data: simpleshapesData, threshold: 0.05 },
     { name: "textscaling", data: textscalingData, threshold: 0.05 },
     { name: "multiplot", data: multiplotData, threshold: 0.05 }
@@ -77,11 +77,11 @@ FIGURES_DATA.forEach(figureData => {
             it("should hover line even if mouse is not exactly on line", function () {
                 cy.window().then((win) => {
                     const draw = win.eval('plot_data');
-                    let [canvasMouse, frameMouse, mouseCoords] = draw.projectMouse({"offsetX": 814, "offsetY": 196} as MouseEvent);
+                    let [canvasMouse, frameMouse, mouseCoords] = draw.projectMouse({"offsetX": 809, "offsetY": 196} as MouseEvent);
                     draw.castMouseMove(canvasMouse, frameMouse, mouseCoords);
                     expect(draw.relativeObjects.shapes[23].isHovered).to.be.true;
 
-                    [canvasMouse, frameMouse, mouseCoords] = draw.projectMouse({"offsetX": 822, "offsetY": 196} as MouseEvent);
+                    [canvasMouse, frameMouse, mouseCoords] = draw.projectMouse({"offsetX": 816, "offsetY": 196} as MouseEvent);
                     draw.castMouseMove(canvasMouse, frameMouse, mouseCoords);
                     expect(draw.relativeObjects.shapes[23].isHovered).to.be.true;
                 });