+++ plot_data/templates.py
@@ -28,36 +28,63 @@ HTML_TEMPLATE = Template('''
       background: #04AA6D;
       cursor: pointer;
     }
+
+    #buttons {
+      display: flex;
+      column-gap: 3rem;
+      row-gap: 0.5em;
+      flex-wrap: wrap;
+      margin-bottom: 0.5rem;
+      margin-top: 1rem;
+      margin: 0 auto;
+    }
+
+    #sub_button {
+      display: flex;
+      gap: 0.5rem;
+      flex-wrap: wrap;
+    }
+
     </style>
     <head>
         <script src=$core_path></script>
     </head>
     <div id="buttons">
-        <button name="mergeON" value="OK" type="button"
-        onclick="plot_data.switchMerge()"> Switch Point Merge </button> &nbsp;&nbsp;
-        <button name="Zoom" value="OK" type="button"
-        onclick="plot_data.switchZoom()"> Zoom Box </button>
-        <button name="Zoom+" value="OK" type="button" onclick="plot_data.zoomIn()"> Zoom+ </button>
-        <button name="Zoom-" value="OK" type="button" onclick="plot_data.zoomOut()"> Zoom- </button> &nbsp;&nbsp;
-        Cluster:&nbsp; <input type="range" class="slider" min="0" max="2500" value="1250"
-        onclick="plot_data.simpleCluster(value / 10000)"></input>
-        <button name="resetClusters" value="OK" type="button"
-        onclick="plot_data.resetClusters()"> Reset clusters </button> &nbsp;&nbsp;
-        <button name="resetView" value="OK" type="button"
-        onclick="plot_data.resetView()"> Reset view </button> &nbsp;&nbsp;
-        <button name="showPoints" value="OK" type="button"
-        onclick="plot_data.togglePoints()"> Show points </button>
-        <button name="switchOrientation" value="OK" type="button"
-        onclick="plot_data.switchOrientation()"> Change Disposition </button>
-        <button name="toogleAxes" value="OK" type="button"
-        onclick="plot_data.htmlToggleAxes()"> Show / Hide Axes </button>
-        <button name="logScale" value="OK" type="button"
-        onclick="plot_data.switchLogScale()"> Log Scale</button>
+        <div id="sub_button">
+            <button name="mergeON" value="OK" type="button"
+            onclick="plot_data.switchMerge()"> Switch Point Merge </button>
+        </div>
+        <div id="sub_button">
+            <button name="Zoom" value="OK" type="button"
+            onclick="plot_data.switchZoom()"> Zoom Box </button>
+            <button name="Zoom+" value="OK" type="button" onclick="plot_data.zoomIn()"> Zoom+ </button>
+            <button name="Zoom-" value="OK" type="button" onclick="plot_data.zoomOut()"> Zoom- </button>
+        </div>
+        <div id="sub_button">
+            Cluster: <input type="range" class="slider" min="0" max="2500" value="1250"
+            onclick="plot_data.simpleCluster(value / 10000)"/>
+            <button name="resetClusters" value="OK" type="button"
+            onclick="plot_data.resetClusters()"> Reset clusters </button>
+        </div>
+        <div id="sub_button">
+            <button name="resetView" value="OK" type="button"
+            onclick="plot_data.resetView()"> Reset view </button>
+            <button name="showPoints" value="OK" type="button"
+            onclick="plot_data.togglePoints()"> Show points </button>
+            <button name="switchOrientation" value="OK" type="button"
+            onclick="plot_data.switchOrientation()"> Change Disposition </button>
+            <button name="toogleAxes" value="OK" type="button"
+            onclick="plot_data.htmlToggleAxes()"> Show / Hide Axes </button>
+            <button name="logScale" value="OK" type="button"
+            onclick="plot_data.switchLogScale()"> Log Scale</button>
+            <button name="resize" value="OK" type="button"
+            onclick="plot_data.resizeWindow(...PlotData.computeCanvasSize('#buttons'))"> Resize </button>
+        </div>
         $specific_buttons
-    <hr style="border-top: 2px;">
     </div>
+    <hr style="border-top: 2px;"/>
     <div id="app">
-        <canvas id="$canvas_id" style="border: 1px solid black;"></canvas>
+        <canvas id="$canvas_id" style="border: 1px solid black; margin: auto;"></canvas>
 
         <script type="text/javascript">
             const [width, height] = PlotData.computeCanvasSize("#buttons");
@@ -71,9 +98,11 @@ HTML_TEMPLATE = Template('''
 EMPTY_BUTTONS = """"""
 
 MULTIPLOT_BUTTONS = """
-        <button name="resize" value="OK" type="button" onclick="plot_data.switchResize()"> Resize Figures </button>
-        <button name="resizeMP" value="OK" type="button"
-        onclick="plot_data.resize(...PlotData.computeCanvasSize('#buttons'))"> Resize Multiplot </button>"""
+        <div id="sub_button">
+            <button name="resize" value="OK" type="button" onclick="plot_data.switchResize()"> Resize Figures </button>
+            <button name="resizeMP" value="OK" type="button"
+            onclick="plot_data.resize(...PlotData.computeCanvasSize('#buttons'))"> Resize Multiplot </button>
+        </div>"""
 
 SCATTER_COMMANDS = """
             const plot_data = new PlotData.Scatter(data, width, height, 0, 0, $canvas_id.id);