+++ src/remoteFigure.ts
@@ -79,6 +79,7 @@ export class RemoteFigure extends Rect {
       this.fixedObjects = new ShapeCollection(this.axes);
       this.relativeObjects = new GroupCollection();
       this.absoluteObjects = new GroupCollection();
+      this.setAxisVisibility(data);
     }
 
   get scale(): Vertex { return new Vertex(this.relativeMatrix.a, this.relativeMatrix.d)}
@@ -116,6 +117,8 @@ export class RemoteFigure extends Rect {
   }
 
   protected unpackData(data: any): Map<string, any[]> { return RemoteFigure.deserializeData(data) }
+
+  protected setAxisVisibility(data: DataInterface): void { this.axes.forEach(axis => axis.visible = data.axis_on) }
 
   public serializeFeatures(): any {
     const elements = [];
@@ -285,6 +288,17 @@ export class RemoteFigure extends Rect {
     this.height = height;
   }
 
+  public changeCanvasSize(width: number, height: number): void {
+    const canvas = document.getElementById(this.canvasID) as HTMLCanvasElement;
+    canvas.width = width;
+    canvas.height = height;
+  }
+
+  public resizeWindow(width: number, height: number): void {
+    this.changeCanvasSize(width, height);
+    this.boundingBoxResize(this.origin, width, height);
+  }
+
   public boundingBoxResize(origin: Vertex, width: number, height: number): void {
     this.changeLocationInCanvas(origin, width, height);
     this.resizeUpdate();
@@ -428,8 +442,10 @@ export class RemoteFigure extends Rect {
 
   public togglePoints(): void {}
 
+  private toggleAxesVisibility(): void { this.axes.forEach(axis => axis.toggleView()) }
+
   public toggleAxes(): void {
-    this.axes.forEach(axis => axis.toggleView());
+    this.toggleAxesVisibility();
     this.draw();
   }
 
@@ -515,6 +531,14 @@ export class RemoteFigure extends Rect {
     this.fixedObjects.mouseUp(ctrlKey);
   }
 
+  public mouseLeave(): void {
+    if (!this.isSelecting && !this.is_drawing_rubber_band && this.translation.normL1 < 10) {
+      this.absoluteObjects.mouseLeave();
+      this.relativeObjects.mouseLeave();
+    }
+    this.fixedObjects.mouseLeave();
+  }
+
   public mouseMoveDrawer(canvas: HTMLElement, e: MouseEvent, canvasDown: Vertex, frameDown: Vertex, clickedObject: Shape): [Vertex, Vertex, Vertex] {
     const [canvasMouse, frameMouse, absoluteMouse] = this.projectMouse(e);
     this.isHovered = this.isInCanvas(absoluteMouse);
@@ -540,6 +564,19 @@ export class RemoteFigure extends Rect {
     return [canvasDown, frameDown, clickedObject]
   }
 
+  private updateWithZoomBox(): void {
+    if (this.isZooming) {
+      if (this.zoomBox.area != 0) this.zoomBoxUpdateAxes(this.zoomBox);
+      this.zoomBox.update(new Vertex(0, 0), new Vertex(0, 0));
+    }
+  }
+
+  private mouseDropRedraw(): [Shape, Vertex] {
+    this.updateWithZoomBox();
+    this.draw();
+    return this.resetMouseEvents()
+  }
+
   public mouseUpDrawer(ctrlKey: boolean): [Shape, Vertex] {
     if (this.isZooming) {
       if (this.zoomBox.area != 0) this.zoomBoxUpdateAxes(this.zoomBox);
@@ -557,7 +594,8 @@ export class RemoteFigure extends Rect {
 
   public mouseLeaveDrawer(canvas: HTMLElement, shiftKey: boolean): [boolean, Vertex] {
     const isZooming = this.isZooming; // TODO: get rid of this with a mousehandler refactor
-    this.mouseUpDrawer(true);
+    this.mouseLeave();
+    this.mouseDropRedraw();
     this.isZooming = isZooming;
     this.axes.forEach(axis => {
       axis.saveLocation();